+++ experimental/packages/opentelemetry-propagator-env-carrier/src/env-carrier.ts
